+++ plugins/inputs/ping/ping_windows_test.go
@@ -163,9 +163,9 @@ func TestLossyPingGather(t *testing.T) {
 		"reply_received":      7,
 		"percent_packet_loss": 22.22222222222222,
 		"percent_reply_loss":  22.22222222222222,
-		"average_response_ms": 115,
-		"minimum_response_ms": 114,
-		"maximum_response_ms": 119,
+		"average_response_ms": 115.0,
+		"minimum_response_ms": 114.0,
+		"maximum_response_ms": 119.0,
 	}
 	acc.AssertContainsTaggedFields(t, "ping", fields, tags)
 }